requirements.txt
# NHTSA Recall & TSB Analyzer
# Core dependencies

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Vector Store
faiss-cpu>=1.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=24.0.0
ruff>=0.1.0
